# portal/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Omni Runner Portal API",
    "description": "API do portal de assessorias esportivas Omni Runner. Gerencia atletas, treinos, financeiro, OmniCoins e administração da plataforma.",
    "version": "1.0.0",
    "contact": {
      "name": "Omni Runner"
    }
  },
  "servers": [
    {
      "url": "https://portal.omnirunner.app",
      "description": "Produção"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/platform/products": {
      "post": {
        "tags": ["Platform Admin"],
        "summary": "Produtos da plataforma",
        "description": "Criar, atualizar, ativar/desativar ou excluir produto de billing. Apenas admin da plataforma.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "required": ["action", "name", "credits_amount", "price_cents"],
                    "properties": {
                      "action": { "type": "string", "enum": ["create"] },
                      "name": { "type": "string", "minLength": 1, "maxLength": 200 },
                      "description": { "type": "string", "maxLength": 1000 },
                      "credits_amount": { "type": "integer", "minimum": 1 },
                      "price_cents": { "type": "integer", "minimum": 1 },
                      "sort_order": { "type": "integer", "minimum": 0 },
                      "product_type": { "type": "string", "enum": ["coins", "badges"] }
                    }
                  },
                  {
                    "type": "object",
                    "required": ["action", "product_id", "is_active"],
                    "properties": {
                      "action": { "type": "string", "enum": ["toggle_active"] },
                      "product_id": { "type": "string", "format": "uuid" },
                      "is_active": { "type": "boolean" }
                    }
                  },
                  {
                    "type": "object",
                    "required": ["action", "product_id"],
                    "properties": {
                      "action": { "type": "string", "enum": ["update"] },
                      "product_id": { "type": "string", "format": "uuid" },
                      "name": { "type": "string" },
                      "description": { "type": "string" },
                      "credits_amount": { "type": "integer" },
                      "price_cents": { "type": "integer" },
                      "sort_order": { "type": "integer" }
                    }
                  },
                  {
                    "type": "object",
                    "required": ["action", "product_id"],
                    "properties": {
                      "action": { "type": "string", "enum": ["delete"] },
                      "product_id": { "type": "string", "format": "uuid" }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string", "enum": ["created", "updated", "deleted"] } }
                }
              }
            }
          },
          "400": { "description": "Validação falhou" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Não é admin da plataforma" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/platform/refunds": {
      "post": {
        "tags": ["Platform Admin"],
        "summary": "Gestão de reembolsos",
        "description": "Aprovar, rejeitar ou processar pedido de reembolso.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action", "refund_id"],
                "properties": {
                  "action": { "type": "string", "enum": ["approve", "reject", "process"] },
                  "refund_id": { "type": "string", "format": "uuid" },
                  "notes": { "type": "string", "maxLength": 1000 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string", "enum": ["approved", "rejected", "processed"] } }
                }
              }
            }
          },
          "400": { "description": "Validação falhou ou reembolso deve ser aprovado antes de processar" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Não é admin da plataforma" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/platform/liga": {
      "post": {
        "tags": ["Platform Admin"],
        "summary": "Gestão da liga",
        "description": "Ações de administração da liga: create_season, activate_season, complete_season, trigger_snapshot.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": { "type": "string", "enum": ["create_season", "activate_season", "complete_season", "trigger_snapshot"] },
                  "name": { "type": "string" },
                  "start_at_ms": { "type": "integer" },
                  "end_at_ms": { "type": "integer" },
                  "season_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "Ação inválida ou campos faltando" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Não é admin da plataforma" },
          "429": { "description": "Rate limit excedido" },
          "503": { "description": "Funcionalidade não disponível" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT do Supabase (access_token da sessão)"
      }
    },
...
  }
}